[evaluation] Extract file-role detection into file_roles module
Move entrypoint-module detection out of characteristic_morphism.py into a
general topos/evaluation/file_roles.py, a home for predicates that classify
a source file's role so evaluation can relax or skip gates (entrypoint is
the first; generated/vendored/test files can follow). Exposes public
is_entrypoint_module.

Address Copilot review on #87:
- Fix language-aware comment handling in _is_entrypoint_source_only: drop the
  global '#' strip that killed C++ #include/#pragma once and Rust #!/#[; allow
  Python #, TS/JS // and #!, and Rust // comments.
- Make is_entrypoint_module a keyword-only flag after threshold in score_simple
  and score_coupling, fixing a latent bug where priority was bound positionally
  into the flag.
- Add regression tests for TS/JS //, C++ header-only, and Rust /// entrypoints.
