+++ topos/evaluation/characteristic_morphism.py
@@ -42,7 +42,6 @@ from __future__ import annotations
 from collections import defaultdict
 from collections.abc import Callable, Iterable, Sequence
 from dataclasses import dataclass, field
-from pathlib import Path
 from typing import TYPE_CHECKING
 
 from topos.core.omega import (
@@ -50,6 +49,7 @@ from topos.core.omega import (
     Omega,
     verdict_from_generators,
 )
+from topos.evaluation.file_roles import is_entrypoint_module
 from topos.evaluation.policies import (
     Priority,
     ScoredDecision,
@@ -137,81 +137,6 @@ _DIMENSION_SCORE_DISPATCHERS: dict[
     "composable": _score_composable_dim,
     "secure": _score_secure_dim,
 }
-
-
-def _entrypoint_filename_hint(path: Path, language: str) -> bool:
-    filename = path.name
-    lower_name = filename.lower()
-    if language == "python":
-        return filename == "__init__.py"
-    if language == "rust":
-        return filename in {"mod.rs", "lib.rs"}
-    if language == "typescript":
-        return lower_name in {"index.ts", "index.tsx"}
-    if language == "javascript":
-        return lower_name in {"index.js", "index.mjs", "index.cjs"}
-    if language == "cpp":
-        return path.suffix.lower() in {".hpp", ".hh", ".hxx"}
-    return False
-
-
-def _is_entrypoint_source_only(source: str, language: str) -> bool:
-    lines = [line.strip() for line in source.splitlines()]
-    lines = [line for line in lines if line and not line.startswith("#")]
-    if not lines:
-        return False
-
-    if language == "python":
-        return all(
-            line.startswith("import ")
-            or line.startswith("from ")
-            or line.startswith("__all__")
-            or line in {"[", "]", "(", ")"}
-            or line.startswith(("'", '"'))
-            for line in lines
-        )
-    if language in {"typescript", "javascript"}:
-        return all(
-            line.startswith("import ")
-            or line.startswith("export *")
-            or line.startswith("export {")
-            or line.startswith("export type ")
-            or line.startswith("export interface ")
-            or line.startswith("/*")
-            or line.startswith("*")
-            or line.endswith("*/")
-            for line in lines
-        )
-    if language == "rust":
-        return all(
-            line.startswith("use ")
-            or line.startswith("pub use ")
-            or line.startswith("pub mod ")
-            or line.startswith("mod ")
-            or line.startswith("extern crate ")
-            or line.startswith("#!")
-            or line.startswith("#[")
-            for line in lines
-        )
-    if language == "cpp":
-        return all(
-            line.startswith("#include")
-            or line.startswith("#pragma once")
-            or line.startswith("//")
-            or line.startswith("/*")
-            or line.startswith("*")
-            or line.endswith("*/")
-            for line in lines
-        )
-    return False
-
-
-def _is_import_export_entrypoint(morphism: ProgramMorphism) -> bool:
-    if morphism.filepath is None:
-        return False
-    if not _entrypoint_filename_hint(morphism.filepath, morphism.language):
-        return False
-    return _is_entrypoint_source_only(morphism.source, morphism.language)
 
 
 # Map each *dimension* name to the singleton generator value it produces
@@ -335,7 +260,7 @@ class CharacteristicMorphism:
         by_dimension: dict[str, list[Representation]] = defaultdict(list)
         for rep in all_reps:
             by_dimension[rep.dimension].append(rep)
-        is_entrypoint_module = _is_import_export_entrypoint(morphism)
+        is_entrypoint = is_entrypoint_module(morphism)
 
         raw_metrics: dict[str, float] = {}
         interpretation: dict[str, str] = {}
@@ -352,7 +277,7 @@ class CharacteristicMorphism:
             if not scorer:
                 continue
 
-            decision = scorer(dim_raw, priority, is_entrypoint_module)
+            decision = scorer(dim_raw, priority, is_entrypoint)
             if decision is None:
                 continue
